requirements.txt
Django==5.1
django-jazzmin==2.6.0
django-crispy-forms==2.1
django-import-export==3.3.1
django-mathfilters==1.0.0
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-taggit==5.0.1
environs==9.5.0
stripe==7.8.1
xhtml2pdf==0.2.11
requests==2.31.0
openpyxl==3.1.2
google-generativeai==0.3.1
Pillow==10.1.0
python-dotenv==1.0.0
